backend/requirements.txt
fastapi
uvicorn
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
openai>=1.0.0
python-multipart>=0.0.6
requests>=2.31.0
aiofiles
azure-data-tables>=12.4.0
azure-storage-queue
pytest
pytest-asyncio
httpx
pytest-cov
kubernetes>=29.0.0
tzdata>=2024.1
PyJWT>=2.8.0
cryptography
